requirements.txt
# M.A.R.K. Rover Control System Dependencies
# Python 3.10+

# Core numerical and scientific computing
numpy>=1.24.0

# Serial communication
pyserial>=3.5

# GNSS/NMEA parsing
pynmea2>=1.18.0

# IMU (BNO085) - Adafruit CircuitPython
adafruit-circuitpython-bno08x>=1.2.0
adafruit-blinka>=8.20.0

# Web API
flask>=2.3.0
flask-cors>=4.0.0
netifaces>=0.11.0  # Network interface detection for IP addresses

# Testing (optional but recommended)
pytest>=7.4.0
